log collection

+++ db/migrate/20150612031900_create_wallet_logs.rb
@@ -5,7 +5,7 @@ class CreateWalletLogs < ActiveRecord::Migration
     create_table :wallet_logs do |t|
       t.references :wallet
       t.integer :action #动作
-      t.integer :balance
+      t.decimal :balance
       t.string :coupons
       t.integer :bean
       t.timestamps null: false
